pyproject.toml
[project]
name = "TATty-agent"
version = "0.1.0"
description = "Agentic RAG Context Engineering - Portable Python package for intelligent code analysis and development"
authors = [
    {name = "TATty Agent Development Team", email = "support@tatty-agent.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["ai", "agent", "code-analysis", "development-tools", "rag", "baml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Console :: Curses",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies (always bundled)
dependencies = [
    # Agent core
    "baml-py>=0.214.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.0.0",

    # File and web operations
    "rich>=13.0.0",
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",

    # Basic data processing
    "pandas>=2.3.3",
    "numpy>=2.2.6",
    "markdown>=3.10",
]

[project.optional-dependencies]
# UI components
tui = ["textual>=0.47.0"]

# Jupyter notebook integration
jupyter = [
    "ipywidgets>=8.0.0",
    "IPython>=8.0.0",
    "jupyter>=1.0.0",
    "notebook>=6.0.0"
]

# Visualization capabilities
visualization = [
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "plotly>=5.0.0"
]

# Enhanced web search
web = [
    "ddgs>=9.9.3",
    "exa-py>=1.16.1"
]

# Document generation and office integration
documents = [
    "weasyprint>=67.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.0",
    "fpdf2>=2.7.0",
    "reportlab>=4.0.0"
]

# Development tools
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0"
]

# Full installation with all features
full = [
    "TATty-agent[tui,jupyter,visualization,web,documents]"
]

[project.urls]
Homepage = "https://github.com/tatty-agent/tatty-agent"
Documentation = "https://docs.tatty-agent.com"
Repository = "https://github.com/tatty-agent/tatty-agent"
"Bug Tracker" = "https://github.com/tatty-agent/tatty-agent/issues"
Changelog = "https://github.com/tatty-agent/tatty-agent/blob/main/CHANGELOG.md"

# CLI entry points
[project.scripts]
tatty-agent = "tatty_agent.cli.main:cli_main"
tatty-init = "tatty_agent.cli.commands:tatty_init"
tatty-tui = "tatty_agent.cli.commands:tatty_tui"
tatty-status = "tatty_agent.cli.commands:tatty_status"

[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

# Tool configurations
[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
]

ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Allow gradual typing

[[tool.mypy.overrides]]
module = [
    "baml_client.*",
    "ddgs.*",
    "exa_py.*",
    "textual.*",
    "weasyprint.*",
    "reportlab.*"
]
ignore_missing_imports = true
